requirements.txt
fastapi
uvicorn
passlib[bcrypt]
pyjwt[crypto]
databases[sqlite]
sqlalchemy
python-multipart
jinja2
pillow
